#6798: avoid double runtime logger initialization
